+++ models/OrderItem.js
@@ -5,7 +5,7 @@ const OrderItem = sequelize.define('OrderItem', {
   id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
   salesOrderId: { type: DataTypes.INTEGER, allowNull: false },
   productId: { type: DataTypes.INTEGER, allowNull: false },
-  quantity: { type: DataTypes.INTEGER, allowNull: false },
+  quantity: { type: DataTypes.DECIMAL(12, 3), allowNull: false },
   unitPrice: { type: DataTypes.DECIMAL(12, 2), defaultValue: 0 },
 }, {
   tableName: 'order_items',
