+++ models/PurchaseOrderItem.js
@@ -7,7 +7,7 @@ const PurchaseOrderItem = sequelize.define('PurchaseOrderItem', {
   productId: { type: DataTypes.INTEGER, allowNull: false },
   productName: { type: DataTypes.STRING, allowNull: true },
   productSku: { type: DataTypes.STRING, allowNull: true },
-  quantity: { type: DataTypes.INTEGER, allowNull: false },
+  quantity: { type: DataTypes.DECIMAL(12, 3), allowNull: false },
   unitPrice: { type: DataTypes.DECIMAL(12, 2), defaultValue: 0 },
   totalPrice: { type: DataTypes.DECIMAL(12, 2), defaultValue: 0 },
 }, {
